WIP: Fix calculation of correlation map
By moving the calculation of the correlation map further back, the u/v
update in the last pass will also be taken into account, making the
correlation map actually reflect the result of all passes.

This does involve a bit more calculation, as image2 to needs to be
interpolated one extra time.